[CK TILE] Fix ambiguous std::abs call for half_t type
Cast to float before calling std::abs to resolve overload ambiguity
with ck_tile::fp16_t/half_t on HIP/ROCm toolchain.

+++ projects/composablekernel/tile_engine/ops/gemm/block_scale_gemm/gemm_aquant/gemm_aquant_benchmark.hpp
@@ -180,10 +180,13 @@ bool compare_aquant(std::string instanceName,
                     ck_tile::HostTensor<CDataType>& c_m_n_dev_result,
                     ck_tile::HostTensor<CDataType>& c_m_n_host_result)
 {
-    const float max_accumulated_value = std::abs(
-        *std::max_element(c_m_n_host_result.mData.begin(),
-                          c_m_n_host_result.mData.end(),
-                          [](const auto& a, const auto& b) { return std::abs(a) < std::abs(b); }));
+    const float max_accumulated_value =
+        std::abs(static_cast<float>(*std::max_element(c_m_n_host_result.mData.begin(),
+                                                      c_m_n_host_result.mData.end(),
+                                                      [](const auto& a, const auto& b) {
+                                                          return std::abs(static_cast<float>(a)) <
+                                                                 std::abs(static_cast<float>(b));
+                                                      })));
     const auto rtol_atol =
         calculate_rtol_atol_aquant<ADataType, AQDataType, BDataType, AccDataType, CDataType>(
             K, kbatch, max_accumulated_value);
